refactor: MemberRole enum 생성 및 Member 필드 추가
