+++ src/main/java/gift/entity/MemberRole.java
@@ -1,0 +1,6 @@
+package gift.entity;
+
+public enum MemberRole {
+    ADMIN,
+    USER
+}
